Pi0.5: address PR review comments
- Refactor Pi05GemmaRMSNorm into shared components/__init__.py (was
  duplicated in paligemma.py). Both paligemma.py and action_expert.py
  now import from the shared location. (sivginirmak)
- Remove dead Pi05Tokenizer.encode_state() method and its unused
  discretize_state import. (merceod)
- Fix sincos_timestep_embedding to compute frequency progression in
  float64 (matching openpi reference) instead of t.dtype which is bf16
  under autocast. (merceod)
- Replace rigid load_pi05_config with auto-mapping: all Pi05Config
  fields are matched by name against HF config keys, with an explicit
  rename dict only for the few keys that differ (num_hidden_layers ->
  num_layers, etc.). (merceod)

All 17 unit tests pass.
